+++ src/nnbench/reporter/__init__.py
@@ -1,0 +1,45 @@
+"""
+A lightweight interface for refining, displaying, and streaming benchmark results to various sinks.
+"""
+from __future__ import annotations
+
+import importlib
+import types
+
+from .base import BenchmarkReporter
+from .console import ConsoleReporter
+
+# internal, mutable
+_reporter_registry: dict[str, type[BenchmarkReporter]] = {
+    "console": ConsoleReporter,
+}
+
+# external, immutable
+reporter_registry: types.MappingProxyType[str, type[BenchmarkReporter]] = types.MappingProxyType(
+    _reporter_registry
+)
+
+
+def register_reporter(key: str, cls_or_name: str | type[BenchmarkReporter]) -> None:
+    """
+    Register a reporter class by its fully qualified module path.
+
+    Parameters
+    ----------
+    key: str
+        The key to register the reporter under. Subsequently, this key can be used in place
+        of reporter classes in code.
+    cls_or_name: str | type[BenchmarkReporter]
+        Name of or full module path to the reporter class. For example, when registering a class
+        ``MyReporter`` located in ``my_module``, ``name`` should be ``my_module.MyReporter``.
+    """
+
+    if isinstance(cls_or_name, str):
+        name = cls_or_name
+        modname, clsname = name.rsplit(".", 1)
+        mod = importlib.import_module(modname)
+        cls = getattr(mod, clsname)
+        _reporter_registry[key] = cls
+    else:
+        # name = cls_or_name.__module__ + "." + cls_or_name.__qualname__
+        _reporter_registry[key] = cls_or_name
